Q0 correction

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -12,7 +12,7 @@ char *_memset(char *s, char b, unsigned int n)
 
 	for (i = 0; s[i] == '\0';)
 		i++;
-	if (i < n)
+	if (i < (int)n)
 		n = i;
 	for (; n > 0; n--)
 		s[n - 1] = b;
